+++ src/paddlefx/__init__.py
@@ -1,14 +1,14 @@
 try:
-    from ._version import version as __version__
-    from ._version import version_tuple
+    from ._version import *  # noqa
 except ImportError:
     __version__ = version = '0.0.0.unknown'
     __version_tuple__ = version_tuple = (0, 0, 0, "unknown")
 
 
-from .graph import Graph
-from .graph_layer import GraphLayer
-from .interpreter import Interpreter as Interpreter
-from .node import Node
-from .proxy import Proxy
-from .symbolic_trace import Tracer, symbolic_trace
+from .eval_frame import DynamoContext, GuardedCode, optimize  # noqa #
+from .graph import Graph  # noqa
+from .graph_layer import GraphLayer  # noqa
+from .interpreter import Interpreter  # noqa
+from .node import Node  # noqa
+from .proxy import Proxy  # noqa
+from .symbolic_trace import Tracer, symbolic_trace  # noqa
